LP-1950 Create Idea and Marketplace App and add all possible pages and their URLs [CHANGE] structural changes as per teams discussion.
