+++ lms/urls.py
@@ -1092,7 +1092,8 @@ if settings.FEATURES.get('ENABLE_API_DOCS'):
 # edx-drf-extensions csrf app
 urlpatterns += [
     url(r'', include('csrf.urls')),
-    url(r'', include('philu.urls')),
+    url(r'', include('openedx.features.marketplace.urls')),
+    url(r'', include('openedx.features.idea.urls')),
 ]
 
 urlpatterns.extend(plugin_urls.get_patterns(plugin_constants.ProjectType.LMS))
