+++ openedx/features/idea/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'openedx.features.idea.apps.IdeaConfig'
