+++ openedx/features/idea/apps.py
@@ -1,10 +1,8 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-
 from django.apps import AppConfig
-from django.conf.urls import url
 
 
 class IdeaConfig(AppConfig):
-    name = 'philu.djangoapps.idea'
+    name = 'openedx.features.idea'
