+++ openedx/features/marketplace/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'openedx.features.marketplace.apps.MarketplaceConfig'
