--- philu/djangoapps/idea/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'philu.djangoapps.idea.apps.IdeaConfig'
