--- philu/djangoapps/idea/views/__init__.py
@@ -1,4 +1,0 @@
-from .challenge_views import ChallengeLandingView
-from .idea_views import IdeaListingView
-from .idea_views import IdeaCreateView
-from .idea_views import IdeaDetailView
