--- philu/djangoapps/marketplace/views/__init__.py
@@ -1,3 +1,0 @@
-from .marketplace_views import MarketplaceListingView
-from .marketplace_views import MarketplaceCreateRequestView
-from .marketplace_views import MarketplaceRequestDetailView
